foodSQUARE Android apk
This is a prototype app which was built for a small project. The aim of this app is to help to save the surplus food from wastage and make it reach the needy. The app is very simple and basic. The app takes updates from places where food is left over and shows them up to the nearby NGO's and volunteers so that they can collect the food and distribute it to the needy.